refactor(contracts): remove single-implementation interfaces
Four interfaces had exactly one implementation each and carried no seam a
downstream project could usefully swap, so they cost a hop through Contracts
without buying anything.

- PermissionAwareColumnInterface was a one-method contract on top of
  ColumnInterface that every column already satisfied. getPermission() moves
  onto ColumnInterface, so ColumnResolver reads it without an instanceof
  probe and a column can no longer silently opt out of permission filtering.
- ColumnAutoDetectorInterface and EditModalTemplateResolverInterface are
  replaced by their concrete classes, un-finaled so test doubles and
  subclasses still work. Their DI aliases go with them; ColumnAutoDetector
  gets a concrete-class alias so autowiring keeps resolving, matching the
  v0.83 treatment of the Mercure and API Platform resolvers.
- LayoutAwareExtensionInterface was a marker with a single implementer.
  ButtonsExtension is the only layout-aware extension, so the two call sites
  test for that class and the reason buttons is skipped by
  DataTableExtensions::jsonSerialize() is documented where it happens. The
  profiler's layoutAware key keeps its name, so the data collector template
  is untouched.

SearchPredicateFactory is merged into DefaultSearchPredicateBuilder: the
factory was a static one-method class the builder forwarded to unchanged.
ContainsSearchStrategy takes the builder as an optional constructor argument
instead of calling the static, so new ContainsSearchStrategy() still works and
a custom builder now reaches the column-control path too.
SearchPredicateBuilderInterface stays: it is a real seam behind
AbstractDataTable::createSearchPredicateBuilder() and the only carrier of the
"must not call andWhere()" contract.

Rejected: a deprecation cycle. The bundle is v0.x, so the removals are direct
and documented in UPGRADE.md instead.

The factory's data-provider test is the real coverage for the type dispatch,
so it moves to DefaultSearchPredicateBuilderTest and replaces the old
delegation-only test.

Confidence: high. Scope risk: low - no payload, route, or template change.

Tested: composer fix, composer audit, vendor/bin/phpunit (1277 tests)
Closes: #382
