+++ config/api_platform.php
@@ -6,7 +6,6 @@ use Pentiminax\UX\DataTables\ApiPlatform\ApiPlatformPropertyTypeMapper;
 use Pentiminax\UX\DataTables\ApiPlatform\ApiResourceCollectionUrlResolver;
 use Pentiminax\UX\DataTables\ApiPlatform\ApiResourceMercureMetadataResolver;
 use Pentiminax\UX\DataTables\ApiPlatform\ColumnAutoDetector;
-use Pentiminax\UX\DataTables\Contracts\ColumnAutoDetectorInterface;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
 use function Symfony\Component\DependencyInjection\Loader\Configurator\service;
@@ -26,7 +25,7 @@ return static function (ContainerConfigurator $container): void {
         ->arg(5, service('datatables.column.property_name_humanizer'))
         ->private();
 
-    $services->alias(ColumnAutoDetectorInterface::class, 'datatables.api_platform.column_auto_detector')
+    $services->alias(ColumnAutoDetector::class, 'datatables.api_platform.column_auto_detector')
         ->private();
 
     $services->set('datatables.api_platform.collection_url_resolver', ApiResourceCollectionUrlResolver::class)
