+++ src/Column/AbstractColumn.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Pentiminax\UX\DataTables\Column;
 
 use Pentiminax\UX\DataTables\Contracts\ColumnInterface;
-use Pentiminax\UX\DataTables\Contracts\PermissionAwareColumnInterface;
 use Pentiminax\UX\DataTables\Enum\ColumnType;
 
 /**
@@ -20,7 +19,7 @@ use Pentiminax\UX\DataTables\Enum\ColumnType;
  * column — see the "Columns" documentation for usage. Only createWithType() below is genuinely
  * internal, restricted to how the bundled column types build themselves.
  */
-abstract class AbstractColumn implements ColumnInterface, PermissionAwareColumnInterface
+abstract class AbstractColumn implements ColumnInterface
 {
     protected ColumnType $type;
     protected ?string $cellType          = null;
