+++ src/Model/Extensions/ButtonsExtension.php
@@ -4,10 +4,9 @@ declare(strict_types=1);
 
 namespace Pentiminax\UX\DataTables\Model\Extensions;
 
-use Pentiminax\UX\DataTables\Contracts\LayoutAwareExtensionInterface;
 use Pentiminax\UX\DataTables\Enum\ButtonType;
 
-final class ButtonsExtension extends AbstractExtension implements LayoutAwareExtensionInterface
+final class ButtonsExtension extends AbstractExtension
 {
     /** @var Button[] */
     private array $buttons = [];
