+++ src/Query/Strategy/ContainsSearchStrategy.php
@@ -6,9 +6,10 @@ namespace Pentiminax\UX\DataTables\Query\Strategy;
 
 use Doctrine\ORM\QueryBuilder;
 use Pentiminax\UX\DataTables\Contracts\ColumnInterface;
+use Pentiminax\UX\DataTables\Contracts\SearchPredicateBuilderInterface;
 use Pentiminax\UX\DataTables\Contracts\SearchStrategyInterface;
 use Pentiminax\UX\DataTables\DataTableRequest\ColumnControlSearch;
-use Pentiminax\UX\DataTables\Query\SearchPredicateFactory;
+use Pentiminax\UX\DataTables\Query\DefaultSearchPredicateBuilder;
 
 /**
  * Strategy for 'contains' search logic.
@@ -16,7 +17,7 @@ use Pentiminax\UX\DataTables\Query\SearchPredicateFactory;
  * Performs a case-sensitive substring search using SQL LIKE %value%.
  * For numeric columns, performs exact match if the value is numeric.
  *
- * Predicate construction is delegated to {@see SearchPredicateFactory} so the
+ * Predicate construction is delegated to {@see SearchPredicateBuilderInterface} so the
  * "numeric → exact / date → skip / text → LIKE" branching lives in a single place.
  * In addition to the column's own type, a search type hint of number/numeric/num
  * forces numeric handling.
@@ -24,6 +25,11 @@ use Pentiminax\UX\DataTables\Query\SearchPredicateFactory;
 final class ContainsSearchStrategy implements SearchStrategyInterface
 {
     private const array NUMERIC_TYPE_HINTS = ['number', 'numeric', 'num'];
+
+    public function __construct(
+        private readonly SearchPredicateBuilderInterface $predicateBuilder = new DefaultSearchPredicateBuilder(),
+    ) {
+    }
 
     public function apply(QueryBuilder $qb, ColumnInterface $column, ColumnControlSearch $search, int $paramIndex, string $alias): void
     {
@@ -39,7 +45,7 @@ final class ContainsSearchStrategy implements SearchStrategyInterface
         $paramName    = \sprintf('column_control_param_%d', $paramIndex);
         $forceNumeric = \in_array(strtolower($search->type), self::NUMERIC_TYPE_HINTS, true);
 
-        $predicate = SearchPredicateFactory::build($qb, $column, $alias, $field, $search->value, $paramName, $forceNumeric);
+        $predicate = $this->predicateBuilder->build($qb, $column, $alias, $field, $search->value, $paramName, $forceNumeric);
 
         if (null !== $predicate) {
             $qb->andWhere($predicate);
