+++ tests/Unit/Controller/AjaxEditFormSubmitControllerTest.php
@@ -12,7 +12,6 @@ use Pentiminax\UX\DataTables\Ajax\AjaxDataTableRegistry;
 use Pentiminax\UX\DataTables\Ajax\AjaxDataTableTokenManager;
 use Pentiminax\UX\DataTables\Attribute\AsDataTable;
 use Pentiminax\UX\DataTables\Column\TextColumn;
-use Pentiminax\UX\DataTables\Contracts\EditModalTemplateResolverInterface;
 use Pentiminax\UX\DataTables\Controller\AjaxEditFormSubmitController;
 use Pentiminax\UX\DataTables\Dto\AjaxEditFormRequestDto;
 use Pentiminax\UX\DataTables\Exception\InvalidCsrfTokenException;
@@ -20,6 +19,7 @@ use Pentiminax\UX\DataTables\Form\ColumnToFormTypeMapper;
 use Pentiminax\UX\DataTables\Form\EditFormBuilder;
 use Pentiminax\UX\DataTables\Form\EditFormService;
 use Pentiminax\UX\DataTables\Form\EditModalRenderer;
+use Pentiminax\UX\DataTables\Form\EditModalTemplateResolver;
 use Pentiminax\UX\DataTables\Mercure\MercureConfig;
 use Pentiminax\UX\DataTables\Mercure\MercureConfigResolver;
 use Pentiminax\UX\DataTables\Mercure\MercureUpdatePublisher;
@@ -108,7 +108,7 @@ final class AjaxEditFormSubmitControllerTest extends TestCase
         $formFactory->expects($this->never())->method('createBuilder');
         $renderer = $this->createMock(EditModalRenderer::class);
         $renderer->expects($this->never())->method('renderBody');
-        $templateResolver = $this->createMock(EditModalTemplateResolverInterface::class);
+        $templateResolver = $this->createMock(EditModalTemplateResolver::class);
         $templateResolver->expects($this->never())->method('resolveColumns');
 
         $authorizationChecker = $this->createMock(AuthorizationCheckerInterface::class);
@@ -213,7 +213,7 @@ final class AjaxEditFormSubmitControllerTest extends TestCase
                 new EntityLocator($registry),
                 new EditFormBuilder($formFactory, new ColumnToFormTypeMapper()),
                 $renderer,
-                $this->createMock(EditModalTemplateResolverInterface::class),
+                $this->createMock(EditModalTemplateResolver::class),
                 new NullMercurePublisher(),
                 permissionChecker: $this->permissionCheckerGranting(true),
             ),
@@ -332,7 +332,7 @@ final class AjaxEditFormSubmitControllerTest extends TestCase
     /**
      * @param string|null $invalidHtml the body rendered for an invalid form, or null when the form is expected to be valid
      *
-     * @return array{FormFactoryInterface, EditModalRenderer, EditModalTemplateResolverInterface}
+     * @return array{FormFactoryInterface, EditModalRenderer, EditModalTemplateResolver}
      */
     private function createFormCollaborators(FormInterface $form, ?string $invalidHtml = null): array
     {
@@ -352,7 +352,7 @@ final class AjaxEditFormSubmitControllerTest extends TestCase
             ->willReturn($formBuilder);
 
         $renderer         = $this->createMock(EditModalRenderer::class);
-        $templateResolver = $this->createMock(EditModalTemplateResolverInterface::class);
+        $templateResolver = $this->createMock(EditModalTemplateResolver::class);
         $templateResolver->expects($this->once())
             ->method('resolveColumns')
             ->with(AjaxEditFormSubmitControllerDataTable::class)
